Extract HTTP mock support
